31/3/2020

+++ jspservletHUHUHU/src/main/webapp/template/custom.js
@@ -97,27 +97,37 @@ $("#btnAssignBuilding").click(function(e) {
 
 //xoá toà nhà
 $("#btnDeleteBuildings").click(function(e) {
-
     e.preventDefault();
     var data = {};
     var buildingIds = $("#buildingList").find('tbody input[type=checkbox]:checked').map(function() {
         return $(this).val();
     }).get();
+    if (buildingIds.length < 1) {
+        alert("bạn chưa chọn toà nhà");
+        return 0;
+    }
     data['buildingIds'] = buildingIds;
-
-    //call api xoá nhiều toà nhà
-    $.ajax({
-        type: "DELETE",
-        url: "http://localhost:8080/api-building",
-        data: JSON.stringify(data),
-        dataType: "json",
-        contentType: "application/json",
-        success: function(response) {
-            console.log("sucess");
-        },
-        error: function(response) {
-            console.log("fail");
+    var thongbao = confirm("Bạn có chắc chắn muốn xoá ");
+    if (thongbao == true) {
+        //gọi ajax xoá toà nhà
+        $.ajax({
+            type: "DELETE",
+            url: "http://localhost:8080/api-building",
+            data: JSON.stringify(data),
+            dataType: "json",
+            contentType: "application/json",
+            success: function(response) {
+                console.log("sucess");
+            },
+            error: function(response) {
+                console.log("fail");
+            }
+        });
+        var i;
+        for (i = 0; i < buildingIds.length; i++) {
+            document.getElementById(`tr_${buildingIds[i]}`).remove();
         }
-    });
+        //$("#boxa > tr#asddasd").hide(1000)
+    }
 
 });
